Move perft out of search
